test(eval-workflows): 按实际职责迁移适配器与测量测试
