+++ test/eval-hosts/runtime-adapter/adapters/custom/command.test.ts
