+++ test/eval-hosts/runtime-adapter/adapters/shared/trial-workspace.test.ts
