+++ test/eval-workflows/measurement/analysis/assertion-layer-core.test.ts
@@ -1,16 +1,16 @@
 import { describe, expect, it } from 'vitest';
-import { DEFAULT_BOOTSTRAP_SEED } from '../../../src/eval-workflows/analysis/bootstrap.js';
+import { DEFAULT_BOOTSTRAP_SEED } from '../../../../src/eval-workflows/analysis/bootstrap.js';
 import {
   canonicalizeJson,
   schemaIdentityKey,
   verifyAnalysisBundle,
   type RuntimeIdentity,
-} from '../../../src/eval-core/contracts/index.js';
+} from '../../../../src/eval-core/contracts/index.js';
 import {
   prepareEvaluationPlan,
   type AnalysisRuntimeRequirement,
   type PreparationRuntime,
-} from '../../../src/eval-core/compiler/index.js';
+} from '../../../../src/eval-core/compiler/index.js';
 import {
   analyzeEvaluationBundleSource,
   createBuiltinAnalysisSchemaValidators,
@@ -18,54 +18,54 @@ import {
   resolveBuiltinAnalysisRuntime,
   type AnalysisNodeImplementation,
   type AnalysisNodeRunContext,
-} from '../../../src/eval-core/analysis/index.js';
+} from '../../../../src/eval-core/analysis/index.js';
 import {
   executeRunPlanSource,
   InMemoryRuntimeEventSequencer,
   type ExecutionClock,
   type ExecutionExecutor,
-} from '../../../src/eval-core/execution/index.js';
+} from '../../../../src/eval-core/execution/index.js';
 import {
   evaluateExecutionBundleSource,
   type EvaluationEvaluator,
-} from '../../../src/eval-core/evaluation/index.js';
+} from '../../../../src/eval-core/evaluation/index.js';
 import {
   ASSERTION_LAYER_ANALYSIS_IMPLEMENTATION_ID,
   createAssertionLayerAnalysisNodes,
-} from '../../../src/eval-workflows/measurement/analysis/assertion-layer-node.js';
+} from '../../../../src/eval-workflows/measurement/analysis/assertion-layer-node.js';
 import {
   createAssertionLayerParameterSchemaValidators,
   parseAssertionLayerParameters,
   type AssertionLayerCriterionParameter,
-} from '../../../src/eval-workflows/measurement/analysis/assertion-layer-parameters.js';
+} from '../../../../src/eval-workflows/measurement/analysis/assertion-layer-parameters.js';
 import {
   createAssertionLayerTableSchemaValidators,
-} from '../../../src/eval-workflows/measurement/analysis/assertion-layer.js';
+} from '../../../../src/eval-workflows/measurement/analysis/assertion-layer.js';
 import {
   COMPOSITE_ANALYSIS_IMPLEMENTATION_ID,
   createCompositeAnalysisNodes,
-} from '../../../src/eval-workflows/measurement/analysis/composite-node.js';
+} from '../../../../src/eval-workflows/measurement/analysis/composite-node.js';
 import {
   createCompositeParameterSchemaValidators,
-} from '../../../src/eval-workflows/measurement/analysis/composite-parameters.js';
+} from '../../../../src/eval-workflows/measurement/analysis/composite-parameters.js';
 import {
   createCompositeTableSchemaValidators,
-} from '../../../src/eval-workflows/measurement/analysis/composite-table.js';
+} from '../../../../src/eval-workflows/measurement/analysis/composite-table.js';
 import {
   BOOTSTRAP_FAMILY_ANALYSIS_IMPLEMENTATION_ID,
   createBootstrapFamilyAnalysisNodes,
-} from '../../../src/eval-workflows/measurement/analysis/bootstrap-family-node.js';
+} from '../../../../src/eval-workflows/measurement/analysis/bootstrap-family-node.js';
 import {
   createBootstrapFamilyParameterSchemaValidators,
-} from '../../../src/eval-workflows/measurement/analysis/bootstrap-family-parameters.js';
+} from '../../../../src/eval-workflows/measurement/analysis/bootstrap-family-parameters.js';
 import {
   createBootstrapFamilyTableSchemaValidators,
-} from '../../../src/eval-workflows/measurement/analysis/bootstrap-family-table.js';
+} from '../../../../src/eval-workflows/measurement/analysis/bootstrap-family-table.js';
 import {
   testRuntime,
   validDefinition,
   validPolicy,
-} from '../../eval-core/compiler/fixtures.js';
+} from '../../../eval-core/compiler/fixtures.js';
 
 class FakeClock implements ExecutionClock {
   #now = 0;
