+++ test/eval-workflows/measurement/analysis/bootstrap-family-table.test.ts
@@ -4,18 +4,18 @@ import {
   bootstrapDiffCI,
   bootstrapMeanCI,
   bootstrapPairedDiffCI,
-} from '../../../src/eval-workflows/analysis/bootstrap.js';
+} from '../../../../src/eval-workflows/analysis/bootstrap.js';
 import {
   digestCanonicalJson,
   schemaIdentityKey,
-} from '../../../src/eval-core/contracts/index.js';
+} from '../../../../src/eval-core/contracts/index.js';
 import {
   BOOTSTRAP_FAMILY_TABLE_SCHEMA,
   buildBootstrapFamilyTable,
   createBootstrapFamilyTableSchemaValidators,
   parseBootstrapFamilyTableEnvelope,
   type BootstrapObservation,
-} from '../../../src/eval-workflows/measurement/analysis/bootstrap-family-table.js';
+} from '../../../../src/eval-workflows/measurement/analysis/bootstrap-family-table.js';
 import {
   BOOTSTRAP_FAMILY_TABLE_V2_SCHEMA,
   BOOTSTRAP_MONTE_CARLO_METHOD_ID,
@@ -23,10 +23,10 @@ import {
   bootstrapSignificanceEvidence,
   createBootstrapFamilyTableV2SchemaValidators,
   parseBootstrapFamilyTableV2Envelope,
-} from '../../../src/eval-workflows/measurement/analysis/bootstrap-family-table-v2.js';
+} from '../../../../src/eval-workflows/measurement/analysis/bootstrap-family-table-v2.js';
 import {
   type BootstrapFamilyParameters,
-} from '../../../src/eval-workflows/measurement/analysis/bootstrap-family-parameters.js';
+} from '../../../../src/eval-workflows/measurement/analysis/bootstrap-family-parameters.js';
 
 const sampleIds = ['sample-1', 'sample-2', 'sample-3', 'sample-4'];
 const controlScores = [1, 2, 3, 4];
