+++ test/eval-workflows/measurement/analysis/composite-parameters.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from 'vitest';
-import { schemaIdentityKey } from '../../../src/eval-core/contracts/index.js';
+import { schemaIdentityKey } from '../../../../src/eval-core/contracts/index.js';
 import {
   COMPOSITE_PARAMETERS_SCHEMA,
   createCompositeParameterSchemaValidators,
   parseCompositeParameters,
-} from '../../../src/eval-workflows/measurement/analysis/composite-parameters.js';
+} from '../../../../src/eval-workflows/measurement/analysis/composite-parameters.js';
 
 const fact = {
   layerId: 'fact',
