+++ test/eval-workflows/measurement/analysis/composite-table.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { digestCanonicalJson, schemaIdentityKey } from '../../../src/eval-core/contracts/index.js';
+import { digestCanonicalJson, schemaIdentityKey } from '../../../../src/eval-core/contracts/index.js';
 import {
   COMPOSITE_TABLE_SCHEMA,
   COMPOSITE_TABLE_SCHEMA_VERSION,
@@ -10,7 +10,7 @@ import {
   type CompositeGroup,
   type CompositeLayerEntry,
   type CompositeTableValue,
-} from '../../../src/eval-workflows/measurement/analysis/composite-table.js';
+} from '../../../../src/eval-workflows/measurement/analysis/composite-table.js';
 
 const assertionSource = digestCanonicalJson('assertion-source');
 const judgeSource = digestCanonicalJson('judge-source');
