+++ test/eval-workflows/measurement/evaluators/llm-assertions.test.ts
@@ -5,56 +5,54 @@ import type {
   MeasurementPolicy,
   RuntimeIdentity,
   UsageRecord,
-} from '../../../src/eval-core/contracts/index.js';
+} from '../../../../src/eval-core/contracts/index.js';
 import {
   RuntimeIdentitySchema,
   digestCanonicalJson,
-} from '../../../src/eval-core/contracts/index.js';
+} from '../../../../src/eval-core/contracts/index.js';
 import {
   prepareEvaluationPlan,
   type PreparationRuntime,
-} from '../../../src/eval-core/compiler/index.js';
+} from '../../../../src/eval-core/compiler/index.js';
 import {
   executeRunPlanSource,
   InMemoryRuntimeEventSequencer,
   type ExecutionClock,
   type ExecutionExecutor,
-} from '../../../src/eval-core/execution/index.js';
+} from '../../../../src/eval-core/execution/index.js';
 import {
   evaluateExecutionBundle,
   type EvaluationCache,
   type EvaluationCacheEntry,
-} from '../../../src/eval-core/evaluation/index.js';
+} from '../../../../src/eval-core/evaluation/index.js';
 import {
-  createRubricJudgeEvaluatorBindingFactory,
-} from '../../../src/eval-hosts/runtime-adapter/evaluators/rubric-judge.js';
+  createLlmAssertionEvaluatorBindingFactory,
+} from '../../../../src/eval-hosts/runtime-adapter/evaluators/llm-assertion-factory.js';
 import {
-  createRubricJudgeEvaluatorIdentity,
-  createRubricJudgeEvaluatorImplementation,
-  captureLlmJudgeInvocationPort,
-  rubricJudgeInstrument,
-  rubricJudgeInstrumentId,
-  RUBRIC_JUDGE_BINDINGS,
-  RUBRIC_JUDGE_CONTEXT_SCHEMA_VERSION,
-  RUBRIC_JUDGE_EVALUATOR_IMPLEMENTATION_ID,
+  createLlmAssertionEvaluatorIdentity,
+  createLlmAssertionEvaluatorImplementation,
+  llmAssertionInstrument,
+  LLM_ASSERTION_BINDINGS,
+  LLM_ASSERTION_CONTEXT_SCHEMA_VERSION,
+  LLM_ASSERTION_EVALUATOR_IMPLEMENTATION_ID,
+  type LlmAssertionType,
+} from '../../../../src/eval-workflows/measurement/evaluators/llm-assertions.js';
+import {
+  createSameProcessEvaluatorAdapter,
+} from '../../../../src/eval-runtime/adapters/same-process.js';
+import {
   type OmkLlmJudgeInvocationPort,
   type OmkLlmJudgeInvocationRequest,
   type OmkLlmJudgeInvocationResult,
-  type RubricJudgeTracePolicy,
-} from '../../../src/eval-hosts/runtime-adapter/index.js';
+} from '../../../../src/eval-hosts/runtime-adapter/index.js';
 import {
-  createSameProcessEvaluatorAdapter,
-} from '../../../src/eval-runtime/adapters/same-process.js';
-import {
-  buildJudgePrompt,
-  getJudgePromptHash,
-  JUDGE_SYSTEM_PROMPT,
-} from '../../../src/eval-runtime/judges/rubric-prompt.js';
-import { buildJudgeTraceSummary } from '../../../src/eval-runtime/judges/trace-summary.js';
-import { testRuntime, validDefinition, validPolicy } from '../../eval-core/compiler/fixtures.js';
+  getRagJudgePromptHash,
+  getSemanticPromptHash,
+} from '../../../../src/eval-workflows/instruments/prompts/judge-prompts.js';
+import { testRuntime, validDefinition, validPolicy } from '../../../eval-core/compiler/fixtures.js';
 
-const PROVIDER_IMPLEMENTATION_ID = 'test.rubric-provider/v1';
-const METRIC_ID = 'rubric-score';
+const PROVIDER_IMPLEMENTATION_ID = 'test.llm-provider/v1';
+const METRIC_ID = 'llm-assertion-pass';
 
 type InvocationHandler = (
   request: Readonly<OmkLlmJudgeInvocationRequest>,
@@ -100,14 +98,11 @@ class MemoryEvaluationCache implements EvaluationCache {
   }
 }
 
-function providerIdentity(revision = 1): RuntimeIdentity {
+function providerIdentity(): RuntimeIdentity {
   return RuntimeIdentitySchema.parse({
     implementationId: PROVIDER_IMPLEMENTATION_ID,
     version: '1.0.0',
-    fingerprint: digestCanonicalJson({
-      provider: PROVIDER_IMPLEMENTATION_ID,
-      revision,
-    }),
+    fingerprint: digestCanonicalJson({ provider: PROVIDER_IMPLEMENTATION_ID, version: 1 }),
     fingerprintBasis: 'content-derived',
     assuranceLevel: 'verified',
     capabilities: { invocation: 'single-call', cancellation: 'cooperative' },
@@ -127,12 +122,32 @@ function invocationPort(
   });
 }
 
-function evaluatorConfig(input: {
-  lengthDebias: boolean;
-  tracePolicy: RubricJudgeTracePolicy;
-  model?: string;
-}): JsonValue {
-  const instrument = rubricJudgeInstrument(input);
+function criterion(
+  assertionType: LlmAssertionType,
+  threshold = 3,
+  negated = false,
+): JsonValue {
+  let source: Record<string, JsonValue>;
+  if (assertionType === 'faithfulness') {
+    source = { context: 'The answer is grounded in this context.' };
+  } else if (assertionType === 'answer_relevancy') {
+    source = { question: 'What is the answer?' };
+  } else {
+    source = { reference: 'Expected answer' };
+  }
+  return {
+    schemaVersion: LLM_ASSERTION_CONTEXT_SCHEMA_VERSION,
+    criterionId: `${assertionType.replaceAll('_', '-')}-criterion`,
+    assertionType,
+    threshold,
+    weight: 1,
+    negated,
+    ...source,
+  };
+}
+
+function evaluatorConfig(assertionType: LlmAssertionType): JsonValue {
+  const instrument = llmAssertionInstrument(assertionType);
   return {
     evaluator: {
       classification: 'public',
@@ -140,75 +155,61 @@ function evaluatorConfig(input: {
     },
     runtime: {
       executorId: PROVIDER_IMPLEMENTATION_ID,
-      model: input.model ?? 'judge-model',
+      model: 'judge-model',
       effort: 'low',
       promptVariant: instrument.promptId,
     },
   } as JsonValue;
 }
 
-function definition(input: {
-  lengthDebias?: boolean;
-  tracePolicy?: RubricJudgeTracePolicy;
-} = {}): EvaluationDefinition {
-  const lengthDebias = input.lengthDebias ?? true;
-  const tracePolicy = input.tracePolicy ?? 'none';
+function definition(
+  assertionType: LlmAssertionType,
+  negated = false,
+): EvaluationDefinition {
   const value = validDefinition();
   value.dataset.samples[0].evaluationContext = {
-    rubricJudge: {
-      schemaVersion: RUBRIC_JUDGE_CONTEXT_SCHEMA_VERSION,
-      criterionId: 'correctness',
-      prompt: 'Answer the question.',
-      rubric: 'The answer must be correct and concise.',
-    },
+    llmAssertion: criterion(assertionType, 3, negated),
   };
-  const instrument = rubricJudgeInstrument({ lengthDebias, tracePolicy });
   value.evaluators = [{
-    evaluatorId: 'rubric-judge',
-    evaluatorKind: 'llm-rubric',
-    implementationId: RUBRIC_JUDGE_EVALUATOR_IMPLEMENTATION_ID,
+    evaluatorId: 'llm-assertion',
+    evaluatorKind: assertionType,
+    implementationId: LLM_ASSERTION_EVALUATOR_IMPLEMENTATION_ID,
     measurement: {
-      instrumentId: rubricJudgeInstrumentId(instrument),
+      instrumentId: llmAssertionInstrument(assertionType).promptId,
       ensembleMemberId: 'provider-member',
-      replicateGroupId: 'rubric-primary',
+      replicateGroupId: `${assertionType.replaceAll('_', '-')}-primary`,
       replicateIndex: 0,
     },
     metricIds: [METRIC_ID],
     inputs: [
-      { bindingId: RUBRIC_JUDGE_BINDINGS.actual, sourceKind: 'output', pointer: '/answer' },
+      { bindingId: LLM_ASSERTION_BINDINGS.actual, sourceKind: 'output', pointer: '/answer' },
       {
-        bindingId: RUBRIC_JUDGE_BINDINGS.criterion,
+        bindingId: LLM_ASSERTION_BINDINGS.criterion,
         sourceKind: 'evaluation-context',
-        pointer: '/rubricJudge',
+        pointer: '/llmAssertion',
       },
-      ...(tracePolicy === 'source-neutral' ? [{
-        bindingId: RUBRIC_JUDGE_BINDINGS.trace,
-        sourceKind: 'trace' as const,
-        pointer: '',
-      }] : []),
     ],
-    config: evaluatorConfig({ lengthDebias, tracePolicy }),
+    config: evaluatorConfig(assertionType),
   }];
   value.metrics = [{
     metricId: METRIC_ID,
-    valueType: 'numeric',
+    valueType: 'boolean',
     scope: 'sample',
-    scale: { min: 1, max: 5 },
     direction: 'higher-is-better',
     missingPolicyId: 'exclude/v1',
   }];
   value.analysisGraph.nodes = [{
     analysisNodeKind: 'reducer',
-    nodeId: 'rubric-mean',
-    implementationId: 'descriptive.mean/v1',
+    nodeId: 'llm-assertion-rate',
+    implementationId: 'descriptive.rate/v1',
     inputs: [{ inputKind: 'metric-observations', referenceId: METRIC_ID }],
-    outputResultId: 'rubric-mean',
+    outputResultId: 'llm-assertion-rate',
   }];
   value.comparisons[0].metricIds = [METRIC_ID];
   value.decisionPolicy = {
     decisionPolicyId: 'release-gate',
     implementationId: 'progress/v1',
-    analysisResultIds: ['rubric-mean'],
+    analysisResultIds: ['llm-assertion-rate'],
     minimumEvidenceStatus: 'complete',
     parameters: { threshold: 0 },
   };
@@ -219,7 +220,6 @@ function policy(input: {
   timeout?: boolean;
   evaluationInvocations?: number;
   retryProviderFailure?: boolean;
-  tracePolicy?: RubricJudgeTracePolicy;
 } = {}): MeasurementPolicy {
   const value = validPolicy();
   value.retry.maxAttempts = 1;
@@ -229,11 +229,14 @@ function policy(input: {
   if (input.retryProviderFailure === true) {
     value.evaluation.retry.maxAttempts = 2;
     value.evaluation.retry.retryableErrorCodes = ['judge-provider-failure'];
-    value.evaluation.retry.backoff = { backoffKind: 'none', initialDelayMs: 0 };
+    value.evaluation.retry.backoff = {
+      backoffKind: 'none',
+      initialDelayMs: 0,
+    };
   }
   if (input.timeout === true) value.evaluation.timeoutMs = 1;
   else delete value.evaluation.timeoutMs;
-  value.evidence.trace = input.tracePolicy === 'source-neutral' ? 'full' : 'none';
+  value.evidence.trace = 'none';
   value.evidence.evidence = 'full';
   value.evidence.maximumClassification = 'gold';
   if (input.evaluationInvocations !== undefined) {
@@ -243,15 +246,11 @@ function policy(input: {
 }
 
 function evaluatorIdentity(
-  input: {
-    lengthDebias: boolean;
-    tracePolicy: RubricJudgeTracePolicy;
-    model?: string;
-  },
+  assertionType: LlmAssertionType,
   port: OmkLlmJudgeInvocationPort,
 ): RuntimeIdentity {
-  const config = evaluatorConfig(input) as unknown as {
-    evaluator: { value: ReturnType<typeof rubricJudgeInstrument> };
+  const config = evaluatorConfig(assertionType) as unknown as {
+    evaluator: { value: ReturnType<typeof llmAssertionInstrument> };
     runtime: {
       executorId: string;
       model: string;
@@ -259,7 +258,7 @@ function evaluatorIdentity(
       promptVariant: string;
     };
   };
-  return createRubricJudgeEvaluatorIdentity({
+  return createLlmAssertionEvaluatorIdentity({
     instrument: config.evaluator.value,
     runtime: config.runtime,
     invocation: port,
@@ -267,10 +266,7 @@ function evaluatorIdentity(
 }
 
 function preparationRuntime(identity: RuntimeIdentity): PreparationRuntime {
-  const base = testRuntime({
-    traceCapability: 'optional',
-    analysisValueTypes: ['numeric'],
-  });
+  const base = testRuntime();
   return {
     ...base,
     resolveEvaluator() {
@@ -294,7 +290,6 @@ function runtimeIdentity(
 
 function executor(
   plan: Awaited<ReturnType<typeof prepareEvaluationPlan>>,
-  tracePolicy: RubricJudgeTracePolicy,
 ): ExecutionExecutor {
   return {
     identity: runtimeIdentity(plan, 'executor', 'control'),
@@ -308,26 +303,6 @@ function executor(
                   value: { answer: 'Actual answer' },
                   classification: 'public' as const,
                 },
-                ...(tracePolicy === 'source-neutral' ? {
-                  trace: {
-                    value: {
-                      schemaVersion: 'omk.source-neutral-trace/v2',
-                      turns: [{ role: 'assistant', content: 'Used search and answered.' }],
-                      toolCalls: [{
-                        tool: 'search',
-                        input: { query: 'Q' },
-                        output: 'A',
-                        success: true,
-                        status: 'success',
-                        statusSource: 'runtime',
-                      }],
-                      numTurns: 1,
-                      fullNumTurns: 1,
-                      numSubAgents: 0,
-                    },
-                    classification: 'sensitive' as const,
-                  },
-                } : {}),
               };
             },
             dispose() {},
@@ -340,22 +315,21 @@ function executor(
 }
 
 async function runCore(input: {
+  assertionType?: LlmAssertionType;
+  negated?: boolean;
   handler: InvocationHandler;
-  lengthDebias?: boolean;
-  tracePolicy?: RubricJudgeTracePolicy;
   policy?: MeasurementPolicy;
   clock?: TestClock;
   signal?: AbortSignal;
   reporting?: 'unsupported' | 'optional' | 'required';
   cache?: EvaluationCache;
 }) {
-  const lengthDebias = input.lengthDebias ?? true;
-  const tracePolicy = input.tracePolicy ?? 'none';
+  const assertionType = input.assertionType ?? 'semantic_similarity';
   const provider = invocationPort(input.handler, input.reporting);
-  const identity = evaluatorIdentity({ lengthDebias, tracePolicy }, provider);
+  const identity = evaluatorIdentity(assertionType, provider);
   const sealed = await prepareEvaluationPlan(
-    definition({ lengthDebias, tracePolicy }),
-    input.policy ?? policy({ tracePolicy }),
+    definition(assertionType, input.negated ?? false),
+    input.policy ?? policy(),
     preparationRuntime(identity),
   );
   const clock = input.clock ?? new TestClock();
@@ -363,22 +337,22 @@ async function runCore(input: {
   const execution = await executeRunPlanSource(sealed, {
     executorsByTargetId: new Map(sealed.execution.targets.map((target) => [
       target.targetId,
-      executor(sealed, tracePolicy),
+      executor(sealed),
     ])),
     clock,
     eventSequencer,
-  }, { runId: 'rubric-judge-run', bundleId: 'rubric-judge-execution' });
+  }, { runId: 'llm-assertion-run', bundleId: 'llm-assertion-execution' });
   const evaluator = createSameProcessEvaluatorAdapter({
     identity,
-    sessionIsolationKey: 'rubric-judge-session',
+    sessionIsolationKey: 'llm-assertion-session',
     resourceLeases: {
       forRun: () => ({
-        bindingId: 'rubric-judge-binding',
+        bindingId: 'llm-assertion-binding',
         consumerKind: 'evaluator',
         resourcesByResourceId: new Map(),
       }),
     },
-    implementation: createRubricJudgeEvaluatorImplementation(provider),
+    implementation: createLlmAssertionEvaluatorImplementation(provider),
   });
   const lifecycle = { runDisposals: 0, recordDisposals: 0 };
   const instrumentedEvaluator = {
@@ -404,19 +378,21 @@ async function runCore(input: {
     },
   };
   const evaluation = await evaluateExecutionBundle(sealed, execution, {
-    evaluatorsByEvaluatorId: new Map([['rubric-judge', instrumentedEvaluator]]),
+    evaluatorsByEvaluatorId: new Map([['llm-assertion', instrumentedEvaluator]]),
     clock,
     eventSequencer,
     ...(input.cache === undefined ? {} : { cache: input.cache }),
   }, {
-    runId: 'rubric-judge-run',
-    bundleId: 'rubric-judge-evaluation',
+    runId: 'llm-assertion-run',
+    bundleId: 'llm-assertion-evaluation',
     ...(input.signal === undefined ? {} : { signal: input.signal }),
   });
-  return { sealed, evaluation, lifecycle };
+  return { sealed, execution, evaluation, provider, lifecycle };
 }
 
-function completedObservations(result: Awaited<ReturnType<typeof runCore>>) {
+function completedObservations(
+  result: Awaited<ReturnType<typeof runCore>>,
+) {
   return result.evaluation.records.flatMap((record) => (
     record.evaluationStatus === 'completed' ? record.observations : []
   ));
@@ -429,24 +405,52 @@ const COMPLETE_USAGE: UsageRecord = {
   providerCost: { amount: 0.001, currency: 'USD', reportedByProvider: true },
 };
 
-describe('provider-neutral rubric raw-reading Evaluator', () => {
+describe('provider-neutral LLM assertion Evaluator', () => {
+  it('preserves opaque provider provenance through evaluator composition', () => {
+    const opaqueProvider = RuntimeIdentitySchema.parse({
+      implementationId: PROVIDER_IMPLEMENTATION_ID,
+      fingerprint: digestCanonicalJson({ provider: PROVIDER_IMPLEMENTATION_ID, opaque: true }),
+      fingerprintBasis: 'opaque',
+      assuranceLevel: 'unknown',
+      capabilities: { invocation: 'single-call' },
+      implementationManifest: {
+        coverageKind: 'fingerprint-plus-facets',
+        facets: [{
+          facetId: 'provider.deployment',
+          value: { coverage: 'remote-opaque' },
+        }],
+      },
+    });
+    const identity = evaluatorIdentity(
+      'semantic_similarity',
+      invocationPort(async () => ({
+        invocationStatus: 'completed',
+        output: '{"score":5,"reason":"valid"}',
+      }), 'optional', opaqueProvider),
+    );
+
+    expect(identity.fingerprintBasis).toBe('opaque');
+    expect(identity.assuranceLevel).toBe('unknown');
+  });
+
   it.each([
-    [true, 'rubric-judge-debias-on', getJudgePromptHash(true), 'v5-cot-toolargs-fmt-len'],
-    [false, 'rubric-judge-debias-off', getJudgePromptHash(false), 'v5-cot-toolargs-fmt'],
-  ] as const)('uses the frozen lengthDebias=%s instrument', async (
-    lengthDebias,
+    ['semantic_similarity', 'semantic-similarity', getSemanticPromptHash()],
+    ['faithfulness', 'rag-faithfulness', getRagJudgePromptHash('faithfulness')],
+    ['answer_relevancy', 'rag-answer-relevancy', getRagJudgePromptHash('answer_relevancy')],
+    ['context_recall', 'rag-context-recall', getRagJudgePromptHash('context_recall')],
+  ] as const)('uses the frozen %s instrument without copying its prompt', async (
+    assertionType,
     promptId,
     promptHash,
-    promptVersion,
   ) => {
     const requests: OmkLlmJudgeInvocationRequest[] = [];
     const result = await runCore({
-      lengthDebias,
+      assertionType,
       handler: async (request) => {
         requests.push(request);
         return {
           invocationStatus: 'completed',
-          output: '{"reasoning":"checked rubric","score":4,"reason":"valid reading"}',
+          output: '{"score":4,"reason":"valid reading"}',
           usage: COMPLETE_USAGE,
         };
       },
@@ -460,98 +464,136 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
       promptId,
       promptHash,
     });
-    expect(requests[0].prompt).toContain(`template ${promptVersion}`);
-    expect(requests[0].prompt).toContain('The answer must be correct and concise.');
-    expect(requests[0].system).toBe(JUDGE_SYSTEM_PROMPT);
-    expect(requests[0].prompt).toBe(buildJudgePrompt(
-      'Answer the question.',
-      'The answer must be correct and concise.',
-      'Actual answer',
-      null,
-      lengthDebias,
-    ));
     expect(requests[0].signal).toBeInstanceOf(AbortSignal);
+    expect(completedObservations(result)).toHaveLength(2);
     for (const observation of completedObservations(result)) {
       expect(observation).toMatchObject({
         metricId: METRIC_ID,
         observationStatus: 'observed',
-        valueType: 'numeric',
-        value: 4,
+        valueType: 'boolean',
+        value: true,
         evidence: {
           classification: 'gold',
           contentKind: 'inline',
           value: {
-            criterionId: 'correctness',
+            assertionType,
             promptId,
             promptHash,
-            lengthDebias,
-            tracePolicy: 'none',
             score: 4,
+            rawPassed: true,
             reason: 'valid reading',
-            reasoning: 'checked rubric',
+            weight: 1,
+            negated: false,
+            layer: 'fact',
           },
         },
       });
     }
+    for (const record of result.evaluation.records) {
+      if (record.evaluationStatus !== 'completed') continue;
+      expect(record.usage).toMatchObject(COMPLETE_USAGE);
+    }
   });
 
-  it('adds only canonical source-neutral trace and keeps the highest input classification', async () => {
-    const requests: OmkLlmJudgeInvocationRequest[] = [];
+  it('keeps a valid below-threshold reading as observed false', async () => {
     const result = await runCore({
-      tracePolicy: 'source-neutral',
-      handler: async (request) => {
-        requests.push(request);
-        return {
-          invocationStatus: 'completed',
-          output: '{"score":5,"reason":"trace supports answer"}',
-        };
-      },
+      handler: async () => ({
+        invocationStatus: 'completed',
+        output: '{"score":2,"reason":"content differs"}',
+      }),
     });
-    expect(requests[0].prompt).toContain('## Agent 执行过程');
-    expect(requests[0].prompt).toContain('search(1)');
-    expect(requests[0].prompt).toContain('Used search and answered.');
-    const expectedTrace = buildJudgeTraceSummary(
-      [{ role: 'assistant', content: 'Used search and answered.' }],
-      [{
-        tool: 'search',
-        input: { query: 'Q' },
-        output: 'A',
-        success: true,
-        status: 'success',
-        statusSource: 'runtime',
-      }],
-    );
-    expect(requests[0].prompt).toBe(buildJudgePrompt(
-      'Answer the question.',
-      'The answer must be correct and concise.',
-      'Actual answer',
-      expectedTrace,
-      true,
-    ));
     expect(completedObservations(result)).toEqual([
+      expect.objectContaining({ observationStatus: 'observed', value: false }),
+      expect.objectContaining({ observationStatus: 'observed', value: false }),
+    ]);
+  });
+
+  it.each([
+    'semantic_similarity',
+    'faithfulness',
+    'answer_relevancy',
+    'context_recall',
+  ] as const)('applies negation only after a valid %s reading', async (assertionType) => {
+    const rawPass = await runCore({
+      assertionType,
+      negated: true,
+      handler: async () => ({
+        invocationStatus: 'completed',
+        output: '{"score":5,"reason":"valid pass"}',
+      }),
+    });
+    expect(completedObservations(rawPass)).toEqual([
       expect.objectContaining({
-        value: 5,
-        evidence: expect.objectContaining({ classification: 'gold' }),
+        observationStatus: 'observed',
+        value: false,
+        evidence: expect.objectContaining({
+          value: expect.objectContaining({ negated: true, rawPassed: true }),
+        }),
       }),
       expect.objectContaining({
-        value: 5,
-        evidence: expect.objectContaining({ classification: 'gold' }),
+        observationStatus: 'observed',
+        value: false,
+        evidence: expect.objectContaining({
+          value: expect.objectContaining({ negated: true, rawPassed: true }),
+        }),
       }),
     ]);
+
+    const rawFail = await runCore({
+      assertionType,
+      negated: true,
+      handler: async () => ({
+        invocationStatus: 'completed',
+        output: '{"score":2,"reason":"valid fail"}',
+      }),
+    });
+    expect(completedObservations(rawFail)).toEqual([
+      expect.objectContaining({
+        observationStatus: 'observed',
+        value: true,
+        evidence: expect.objectContaining({
+          value: expect.objectContaining({ negated: true, rawPassed: false }),
+        }),
+      }),
+      expect.objectContaining({
+        observationStatus: 'observed',
+        value: true,
+        evidence: expect.objectContaining({
+          value: expect.objectContaining({ negated: true, rawPassed: false }),
+        }),
+      }),
+    ]);
+  });
+
+  it('seals negation into the EvaluationPlan identity', async () => {
+    const provider = invocationPort(async () => ({
+      invocationStatus: 'completed',
+      output: '{"score":5,"reason":"valid"}',
+    }));
+    const identity = evaluatorIdentity('semantic_similarity', provider);
+    const positive = await prepareEvaluationPlan(
+      definition('semantic_similarity', false),
+      policy(),
+      preparationRuntime(identity),
+    );
+    const negated = await prepareEvaluationPlan(
+      definition('semantic_similarity', true),
+      policy(),
+      preparationRuntime(identity),
+    );
+    expect(positive.evaluation.evaluationPlanDigest)
+      .not.toBe(negated.evaluation.evaluationPlanDigest);
   });
 
   it.each([
     ['plain text', 'judge-response-non-json'],
     ['prefix {bad json} suffix', 'judge-response-malformed-json'],
-    ['prefix {"score":4,"reason":"embedded"} suffix', 'judge-response-malformed-json'],
     ['{"score":"4","reason":"wrong type"}', 'judge-score-malformed'],
-    ['{"score":4.5,"reason":"fractional"}', 'judge-score-malformed'],
-    ['{"score":0,"reason":"out of range"}', 'judge-score-out-of-range'],
     ['{"score":6,"reason":"out of range"}', 'judge-score-out-of-range'],
     ['{"score":4}', 'judge-reason-missing'],
-    ['{"score":4,"reason":"  "}', 'judge-reason-missing'],
-  ])('returns invalid, not a zero reading, for %s', async (output, reasonCode) => {
+  ])('returns invalid, not false, for %s even when negated', async (output, reasonCode) => {
     const result = await runCore({
+      negated: true,
       handler: async () => ({ invocationStatus: 'completed', output }),
     });
     expect(completedObservations(result)).toEqual([
@@ -563,8 +605,9 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
     ))).toBe(false);
   });
 
-  it('records provider failure separately, redacts provider details, and retains known usage', async () => {
+  it('does not invert provider failure, redacts raw errors, and retains known usage', async () => {
     const result = await runCore({
+      negated: true,
       handler: async () => ({
         invocationStatus: 'failed',
         reasonCode: 'provider-overloaded',
@@ -574,6 +617,7 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
         },
       }),
     });
+    expect(result.evaluation.records).toHaveLength(2);
     for (const record of result.evaluation.records) {
       expect(record).toMatchObject({
         evaluationStatus: 'failed',
@@ -589,7 +633,7 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
     }
   });
 
-  it('keeps unknown provider usage and cost absent', async () => {
+  it('keeps unknown provider usage and cost absent rather than writing zero', async () => {
     const result = await runCore({
       handler: async () => ({
         invocationStatus: 'completed',
@@ -604,14 +648,17 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
     }
   });
 
-  it('delegates retry exclusively to sealed Core policy', async () => {
+  it('delegates retry exclusively to the sealed Core retry policy', async () => {
     let calls = 0;
     const result = await runCore({
       handler: async () => {
         calls += 1;
         return calls <= 2
           ? { invocationStatus: 'failed', reasonCode: 'provider-unavailable' }
-          : { invocationStatus: 'completed', output: '{"score":5,"reason":"recovered"}' };
+          : {
+              invocationStatus: 'completed',
+              output: '{"score":5,"reason":"recovered"}',
+            };
       },
       policy: policy({ retryProviderFailure: true }),
     });
@@ -619,7 +666,10 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
     for (const record of result.evaluation.records) {
       expect(record).toMatchObject({
         evaluationStatus: 'completed',
-        observations: [expect.objectContaining({ value: 5 })],
+        observations: [expect.objectContaining({
+          observationStatus: 'observed',
+          value: true,
+        })],
       });
       if (record.evaluationStatus !== 'completed') continue;
       expect(record.attempts.map((attempt) => attempt.attemptStatus)).toEqual([
@@ -638,7 +688,7 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
       handler: async (): Promise<OmkLlmJudgeInvocationResult> => {
         calls += 1;
         return {
-          invocationStatus: 'completed',
+          invocationStatus: 'completed' as const,
           output: '{"score":5,"reason":"cacheable"}',
         };
       },
@@ -648,7 +698,8 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
     const first = await runCore(request);
     expect(calls).toBe(2);
     expect(first.evaluation.records.every((record) => (
-      record.evaluationStatus === 'completed' && record.cache.cacheStatus === 'miss'
+      record.evaluationStatus === 'completed'
+      && record.cache.cacheStatus === 'miss'
     ))).toBe(true);
 
     const second = await runCore(request);
@@ -659,7 +710,7 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
     ))).toBe(true);
   });
 
-  it('lets Core own timeout and cooperative provider cancellation', async () => {
+  it('keeps negated timeout failed and waits for cooperative provider cancellation', async () => {
     let cancellations = 0;
     const result = await runCore({
       handler: async (request) => new Promise<OmkLlmJudgeInvocationResult>((_resolve, reject) => {
@@ -668,17 +719,20 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
           reject(request.signal.reason);
         }, { once: true });
       }),
+      negated: true,
       policy: policy({ timeout: true }),
       clock: new TestClock(true),
     });
     expect(cancellations).toBe(2);
-    expect(result.evaluation.records.every((record) => (
-      record.evaluationStatus === 'failed'
-      && record.error.code === 'timeout'
-    ))).toBe(true);
+    for (const record of result.evaluation.records) {
+      expect(record).toMatchObject({
+        evaluationStatus: 'failed',
+        error: { code: 'timeout', stage: 'evaluation' },
+      });
+    }
   });
 
-  it('lets Core own cancellation without converting it into provider failure', async () => {
+  it('keeps negated cancellation cancelled instead of turning it into a pass', async () => {
     const controller = new AbortController();
     let cancellations = 0;
     const result = await runCore({
@@ -696,6 +750,7 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
           }
         });
       },
+      negated: true,
       signal: controller.signal,
     });
     expect(cancellations).toBeGreaterThan(0);
@@ -705,7 +760,7 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
     ))).toBe(true);
   });
 
-  it('lets Core censor unstarted coordinates at the sealed evaluation budget', async () => {
+  it('keeps negated budget censoring as unstarted evidence', async () => {
     let calls = 0;
     const result = await runCore({
       handler: async () => {
@@ -715,105 +770,55 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
           output: '{"score":5,"reason":"valid"}',
         };
       },
+      negated: true,
       policy: policy({ evaluationInvocations: 1 }),
     });
     expect(calls).toBe(1);
     expect(result.evaluation.evaluationBundleStatus).toBe('budget-exhausted');
-    expect(result.evaluation.coverage).toMatchObject({ planned: 2, completed: 1, notStarted: 1 });
-  });
-
-  it('fingerprints prompt variant, trace policy, model, and provider Runtime identity', () => {
-    const handler: InvocationHandler = async () => ({
-      invocationStatus: 'completed',
-      output: '{"score":5,"reason":"valid"}',
+    expect(result.evaluation.coverage).toMatchObject({
+      planned: 2,
+      completed: 1,
+      notStarted: 1,
     });
-    const basePort = invocationPort(handler);
-    const base = evaluatorIdentity({ lengthDebias: true, tracePolicy: 'none' }, basePort);
-    const identities = [
-      evaluatorIdentity({ lengthDebias: false, tracePolicy: 'none' }, basePort),
-      evaluatorIdentity({ lengthDebias: true, tracePolicy: 'source-neutral' }, basePort),
-      evaluatorIdentity({ lengthDebias: true, tracePolicy: 'none', model: 'other' }, basePort),
-      evaluatorIdentity(
-        { lengthDebias: true, tracePolicy: 'none' },
-        invocationPort(handler, 'optional', providerIdentity(2)),
-      ),
-    ];
-    for (const identity of identities) expect(identity.fingerprint).not.toBe(base.fingerprint);
   });
 
-  it('preserves opaque provider provenance through evaluator composition', () => {
-    const opaqueProvider = RuntimeIdentitySchema.parse({
-      implementationId: PROVIDER_IMPLEMENTATION_ID,
-      fingerprint: digestCanonicalJson({ provider: PROVIDER_IMPLEMENTATION_ID, opaque: true }),
-      fingerprintBasis: 'opaque',
-      assuranceLevel: 'unknown',
-      capabilities: { invocation: 'single-call' },
-      implementationManifest: {
-        coverageKind: 'fingerprint-plus-facets',
-        facets: [{
-          facetId: 'provider.deployment',
-          value: { coverage: 'remote-opaque' },
-        }],
+  it('cannot lower observed content score by adding an infrastructure failure', async () => {
+    let calls = 0;
+    const result = await runCore({
+      handler: async () => {
+        calls += 1;
+        return calls === 1
+          ? {
+              invocationStatus: 'completed',
+              output: '{"score":5,"reason":"valid"}',
+            }
+          : {
+              invocationStatus: 'failed',
+              reasonCode: 'provider-unavailable',
+            };
       },
     });
-    const identity = evaluatorIdentity(
-      { lengthDebias: true, tracePolicy: 'none' },
-      invocationPort(async () => ({
-        invocationStatus: 'completed',
-        output: '{"score":5,"reason":"valid"}',
-      }), 'optional', opaqueProvider),
-    );
-
-    expect(identity.fingerprintBasis).toBe('opaque');
-    expect(identity.assuranceLevel).toBe('unknown');
+    const observed = completedObservations(result).filter((observation) => (
+      observation.observationStatus === 'observed'
+    ));
+    expect(observed).toHaveLength(1);
+    expect(observed[0]).toMatchObject({ value: true });
+    expect(result.evaluation.records.some((record) => (
+      record.evaluationStatus === 'completed'
+    ))).toBe(true);
+    expect(result.evaluation.records.some((record) => (
+      record.evaluationStatus === 'failed'
+      && record.error.code === 'judge-provider-failure'
+    ))).toBe(true);
   });
 
-  it('captures the provider method and receiver identity against later mutation', async () => {
-    const initialIdentity = providerIdentity(1);
-    let originalCalls = 0;
-    const mutablePort: {
-      identity: RuntimeIdentity;
-      providerCost: { reporting: 'optional' };
-      invoke: OmkLlmJudgeInvocationPort['invoke'];
-    } = {
-      identity: initialIdentity,
-      providerCost: { reporting: 'optional' as const },
-      async invoke(this: OmkLlmJudgeInvocationPort) {
-        originalCalls += 1;
-        expect(this.identity).toEqual(initialIdentity);
-        return {
-          invocationStatus: 'completed' as const,
-          output: '{"score":5,"reason":"captured"}',
-        };
-      },
-    };
-    const captured = captureLlmJudgeInvocationPort(mutablePort);
-    mutablePort.identity = providerIdentity(2);
-    mutablePort.invoke = async () => ({
-      invocationStatus: 'failed' as const,
-      reasonCode: 'mutated-method',
-    });
-    const result = await captured.invoke({
-      executorId: PROVIDER_IMPLEMENTATION_ID,
-      model: 'judge-model',
-      system: 'system',
-      prompt: 'prompt',
-      promptId: 'rubric-judge-debias-on',
-      promptHash: getJudgePromptHash(true),
-      signal: new AbortController().signal,
-    });
-    expect(result).toMatchObject({ invocationStatus: 'completed' });
-    expect(originalCalls).toBe(1);
-    expect(captured.identity).toEqual(initialIdentity);
-  });
-
-  it('builds a host factory with sealed qualification and shared provider boundary', async () => {
-    const definitionValue = definition();
+  it('builds a production host factory with sealed qualification and prompt identity', async () => {
+    const definitionValue = definition('semantic_similarity');
     const provider = invocationPort(async () => ({
       invocationStatus: 'completed',
       output: '{"score":5,"reason":"valid"}',
     }));
-    const factory = createRubricJudgeEvaluatorBindingFactory(async () => ({
+    const factory = createLlmAssertionEvaluatorBindingFactory(async () => ({
       port: provider,
       preflightDeclarations: [{
         preflightKind: 'connectivity',
@@ -823,12 +828,12 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
       }],
     }));
     const resolved = await factory({
-      sessionIsolationKey: 'rubric-judge-factory-session',
+      sessionIsolationKey: 'llm-assertion-factory-session',
       binding: {
         runtimeKind: 'evaluator',
-        bindingId: 'rubric-judge-binding',
-        evaluatorId: 'rubric-judge',
-        implementationId: RUBRIC_JUDGE_EVALUATOR_IMPLEMENTATION_ID,
+        bindingId: 'llm-assertion-binding',
+        evaluatorId: 'llm-assertion',
+        implementationId: LLM_ASSERTION_EVALUATOR_IMPLEMENTATION_ID,
         measurement: definitionValue.evaluators[0].measurement,
         configDigest: digestCanonicalJson(definitionValue.evaluators[0].config as JsonValue),
         resourceLeaseRequirements: [],
@@ -836,23 +841,20 @@ describe('provider-neutral rubric raw-reading Evaluator', () => {
           executorId: PROVIDER_IMPLEMENTATION_ID,
           model: 'judge-model',
           effort: 'low',
-          promptVariant: 'rubric-judge-debias-on',
+          promptVariant: 'semantic-similarity',
           resourceIntegrity: 'digest-before-use',
         },
       },
       evaluator: definitionValue.evaluators[0],
       resourceLeases: {
         forRun: () => ({
-          bindingId: 'rubric-judge-binding',
+          bindingId: 'llm-assertion-binding',
           consumerKind: 'evaluator',
           resourcesByResourceId: new Map(),
         }),
       },
     });
-    expect(resolved.port.identity).toEqual(evaluatorIdentity({
-      lengthDebias: true,
-      tracePolicy: 'none',
-    }, provider));
+    expect(resolved.port.identity).toEqual(evaluatorIdentity('semantic_similarity', provider));
     expect(resolved.preflightDeclarations).toEqual([
       expect.objectContaining({ checkId: 'judge-connectivity', preflightDisposition: 'check' }),
     ]);
